Snow should be handled by a populator, not decorator
